Add UserCreatedModelMixin
